Move remaining DEFAULT_* constants to `const.defaults`
Some imports are made locally to avoid circular import issues (mainly in `util` module)
