+++ picard/file.py
@@ -66,7 +66,7 @@ from picard import (
     log,
 )
 from picard.config import get_config
-from picard.const import DEFAULT_TIME_FORMAT
+from picard.const.defaults import DEFAULT_TIME_FORMAT
 from picard.const.sys import (
     IS_MACOS,
     IS_WIN,
