+++ picard/ui/widgets/scriptlistwidget.py
@@ -30,7 +30,7 @@ from PyQt6 import (
     QtWidgets,
 )
 
-from picard.const import DEFAULT_SCRIPT_NAME
+from picard.const.defaults import DEFAULT_SCRIPT_NAME
 from picard.i18n import (
     gettext as _,
     gettext_constants,
